+++ Lesson17Homework3/Lesson17Homework3/Lesson17Homework3.cpp
@@ -5,8 +5,8 @@ using namespace std;
 int main()
 {
     // User's Age After 5 Years
-    int Age = 25;
-    int YourAgeAfter5Years = Age + 5;
+    short Age = 25;
+    short YourAgeAfter5Years = Age + 5;
 
     cout << "After 5 years you will be " << Age + 5 << " years old." << endl;
     cout << "After 5 years you will be " << YourAgeAfter5Years << " years old." << endl;
