+++ Lesson18Homework3/Lesson18Homework3/Lesson18Homework3.cpp
@@ -5,12 +5,12 @@ using namespace std;
 int main()
 {
     // User's Age After 5 Years
-    int Age;
+    short Age;
 
     cout << "Please enter your age?" << endl;
     cin >> Age;
 
-    int YourAgeAfter5Years = Age + 5;
+    short YourAgeAfter5Years = Age + 5;
 
     cout << "After 5 years you will be " << YourAgeAfter5Years << " years old." << endl;
